refactor: Rename `lib/plugins/aws/package/lib/saveServiceState.js`
