+++ lib/plugins/aws/package/index.js
@@ -5,7 +5,7 @@ const mergeCustomProviderResources = require('./lib/merge-custom-provider-resour
 const stripNullPropsFromTemplateResources = require('./lib/stripNullPropsFromTemplateResources');
 const generateArtifactDirectoryName = require('./lib/generate-artifact-directory-name');
 const generateCoreTemplate = require('./lib/generate-core-template');
-const saveServiceState = require('./lib/saveServiceState');
+const saveServiceState = require('./lib/save-service-state');
 const saveCompiledTemplate = require('./lib/saveCompiledTemplate');
 const mergeIamTemplates = require('./lib/merge-iam-templates');
 const addExportNameForOutputs = require('./lib/add-export-name-for-outputs');
