+++ lib/plugins/aws/package/lib/save-service-state.js
