+++ test/unit/lib/plugins/aws/package/lib/save-service-state.test.js
